Refactoring and adding comments
