+++ composer.json
@@ -60,12 +60,12 @@
         "symfony/yaml" : "^2.8",
         "tdammers/htmlmaid" : "~0.7",
         "twig/twig" : "^1.18",
-        "ua-parser/uap-php" : "~3.4"
+        "ua-parser/uap-php" : "~3.4",
+        "lstrojny/phpunit-function-mocker": "^0.2.0"
     },
     "require-dev" : {
         "bolt/codingstyle" : "~1.0",
         "codeception/codeception" : "^2.1",
-        "lstrojny/phpunit-function-mocker" : "dev-master@dev",
         "phpmd/phpmd" : "~2.2",
         "phpunit/dbunit" : "~1.3",
         "phpunit/php-code-coverage" : "~2.0",
