+++ src/Controller/Backend/Users.php
@@ -1,7 +1,6 @@
 <?php
 namespace Bolt\Controller\Backend;
 
-use Silex\Application;
 use Bolt\AccessControl\Permissions;
 use Bolt\Events\AccessControlEvent;
 use Bolt\Storage\Entity;
@@ -13,7 +12,6 @@ use Symfony\Component\Form\Extension\Core\Type\FormType;
 use Symfony\Component\Form\Extension\Core\Type\HiddenType;
 use Symfony\Component\Form\Extension\Core\Type\PasswordType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
-use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\Form;
 use Symfony\Component\Form\FormBuilder;
 use Symfony\Component\Form\FormError;
@@ -21,24 +19,6 @@ use Symfony\Component\Form\FormEvent;
 use Symfony\Component\Form\FormEvents;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Validator\Constraints as Assert;
-use Symfony\Component\Validator\Constraints\Email as EmailConstraint;
-
-
-use Symfony\Component\HttpKernel\Debug\ErrorHandler;
-use Symfony\Component\HttpKernel\Debug\ExceptionHandler;
-
-// set the error handling
-ini_set('display_errors', 1);
-error_reporting(-1);
-ErrorHandler::register();
-if ('cli' !== php_sapi_name()) {
-    ExceptionHandler::register();
-}
-
-
-// set debug mode
-$app['debug'] = true;
-
 
 /**
  * Backend controller for user maintenance routes.
@@ -59,19 +39,6 @@ class Users extends BackendBase
             ->assert('id', '\d*')
             ->bind('useredit');
 
-        $c->match('/users/invitationlink', 'invitationLink')
-            ->bind('invitationlink');
-
-        $c->match('/users/generatelink', 'generateLink')
-            ->bind('generatelink');
-
-        $c->match('/users/sendlink', 'sendLink')
-            ->bind('sendlink');
-
-        $c->match('/invitation/{code}', 'invitation')
-            ->assert('code', '.*')
-            ->bind('invitation');
-
         $c->match('/userfirst', 'first')
             ->bind('userfirst');
 
@@ -83,6 +50,10 @@ class Users extends BackendBase
 
         $c->get('/roles', 'viewRoles')
             ->bind('roles');
+
+        $c->match('/invitation/{code}', 'invitation')
+            ->assert('code', '.*')
+            ->bind('invitation');
     }
 
     /**
@@ -175,192 +146,6 @@ class Users extends BackendBase
         return $this->render('@bolt/edituser/edituser.twig', $context);
     }
 
-
-    /**
-     * Invitation link route.
-     *
-     * @return \Bolt\Response\BoltResponse|\Symfony\Component\HttpFoundation\RedirectResponse
-     */
-    public function invitationLink()
-    {
-        $currentUser = $this->getUser();
-
-        // Get the base form
-        $form = $this->getGenerateInvitationForm($this->getUser());
-
-        // Generate the form
-        $form = $form->getForm();
-
-        /** @var \Symfony\Component\Form\FormView|\Symfony\Component\Form\FormView[] $formView */
-        $formView = $form->createView();
-
-        // Get the base form
-        $form = $this->getInvitationEmailForm($this->getUser());
-
-        // Generate the form
-        $form = $form->getForm();
-
-        /** @var \Symfony\Component\Form\FormView|\Symfony\Component\Form\FormView[] $formView */
-        $formEmailView = $form->createView();
-
-        $manipulatableRoles = $this->app['permissions']->getManipulatableRoles($currentUser->toArray());
-        foreach ($formView['roles'] as $role) {
-            if (!in_array($role->vars['value'], $manipulatableRoles)) {
-                $role->vars['attr']['disabled'] = 'disabled';
-            }
-        }
-
-        $context = [
-            'form'        => $formView,
-            'emailform'        => $formEmailView,
-        ];
-
-        return $this->render('@bolt/invitation/generate.twig', $context);
-    }
-
-    /**
-     * Invitation link route.
-     *
-     * @param Request $request The Symfony Request
-     *
-     * @return \Bolt\Response\BoltResponse|\Symfony\Component\HttpFoundation\RedirectResponse
-     */
-    public function generateLink(Request $request, Application $app)
-    {
-        $roles = array();
-
-        if(null !== ($request->request->get('roles'))) {
-            $roles = $request->request->get('roles');
-        }
-
-        $expiration_date = $request->request->get('expiration_date');
-        $expiration_time = $request->request->get('expiration_time');
-
-        //validate expiration date (not empty, not in the past)
-        $expiration = new \DateTime(str_replace("/","-",$expiration_date." ".$expiration_time));
-        $expiration->format('Y-m-d H:i:s');
-
-        $book = array(
-            'empty' => $expiration_date,
-            'expire' => $expiration,
-        );
-
-        $constraint = new Assert\Collection(array(
-            'empty' => new Assert\NotBlank(),
-            'expire' => new Assert\Range(array(
-                'min' => 'now',
-                )),
-            )
-        );
-
-        $errors = $app['validator']->validate($book, $constraint);
-
-        if (count($errors) > 0) {
-            foreach ($errors as $error) {
-                return $app->json($error->getMessage(), 400);
-            }
-        }
-
-        $token = bin2hex(openssl_random_pseudo_bytes(16));
-
-        $tokenEntity = new Entity\Tokens();
-
-        $tokenEntity->setToken($token);
-        $tokenEntity->setRoles($roles);
-        $tokenEntity->setExpiration($expiration);
-
-        $this->getRepository('Bolt\Storage\Entity\Tokens')->save($tokenEntity);
-
-        return $token;
-    }
-
-    /**
-     * Invitation link route.
-     *
-     * @param Request $request The Symfony Request
-     *
-     * @return \Bolt\Response\BoltResponse|\Symfony\Component\HttpFoundation\RedirectResponse
-     */
-    public function sendLink(Request $request, Application $app)
-    {
-        $to = $request->request->get('to');
-        $subject = $request->request->get('subject');
-        $text = $request->request->get('message');
-        $link = $request->request->get('link');
-
-        //validate send email form
-        $book = array(
-            'to' => $to,
-            'subject' => $subject,
-            'text' => $text,
-        );
-
-        $constraint = new Assert\Collection(array(
-                'to' => new Assert\Email(),
-                'subject' => new Assert\NotBlank(),
-                'text' => new Assert\NotBlank(),
-            )
-        );
-
-        $violationList = $app['validator']->validate($book, $constraint);
-
-        if (count($violationList) > 0) {
-            foreach ($violationList as $violation){
-                $field = preg_replace('/\[|\]/', "", $violation->getPropertyPath());
-                $error = $violation->getMessage();
-                $errors[$field] = $error;
-            }
-            return $app->json($errors, 400);
-        }
-
-        $userEntity = new Entity\Users($this->getUser());
-        $from = $this->app['config']->get('general/mailoptions/senderMail', $userEntity->getEmail());
-
-        // Compile the email with the invitation link.
-        $mailhtml = $this->app['render']->render(
-            '@bolt/mail/invitation.twig',
-            [
-                'message'           => $text,
-                'link'              => $link,
-            ]
-        );
-
-        $message = $this->app['mailer']
-            ->createMessage('message')
-            ->setSubject($subject)
-            ->setFrom($from)
-            ->setReplyTo($from)
-            ->setTo($to)
-            ->setBody(strip_tags($mailhtml))
-            ->addPart($mailhtml, 'text/html')
-        ;
-
-        $failed = true;
-        $failedRecipients = [];
-
-        try {
-            $recipients = $this->app['mailer']->send($message, $failedRecipients);
-
-            // Try and send immediately
-            $this->app['swiftmailer.spooltransport']->getSpool()->flushQueue($this->app['swiftmailer.transport']);
-
-            if ($recipients) {
-                $this->app['logger.system']->info("Invitation request sent to '" . $to . "'.", ['event' => 'authentication']);
-                $failed = false;
-            }
-        } catch (\Exception $e) {
-            // Notify below
-        }
-
-        if ($failed) {
-            $this->app['logger.system']->error("Failed to send invitation request sent to '" . $to . "'.", ['event' => 'authentication']);
-
-            return false;
-        }
-
-        return true;
-    }
-
     /**
      * Create the first user.
      *
@@ -370,7 +155,6 @@ class Users extends BackendBase
      */
     public function first(Request $request)
     {
-
         // We should only be here for creating the first user
         if ($this->app['schema']->hasUserTable() && $this->users()->hasUsers()) {
             return $this->redirectToRoute('dashboard');
@@ -416,54 +200,6 @@ class Users extends BackendBase
         ];
 
         return $this->render('@bolt/firstuser/firstuser.twig', $context);
-    }
-
-    /**
-     * Create a user from invitation link.
-     *
-     * @param Request $request  The Symfony Request
-     * @param String $code      The invitation code provided
-     *
-     * @return \Bolt\Response\BoltResponse|\Symfony\Component\HttpFoundation\RedirectResponse
-     */
-    public function invitation(Request $request, $code)
-    {
-        // Check if the invitation code exists and it is still available
-        $token = $this->getRepository('Bolt\Storage\Entity\Tokens')->findBy(array('token' => $code));
-
-        if((!$token) || (new \DateTime() > new \DateTime($token[0]->expiration))){
-            return $this->render('@bolt/invitation/invitationfail.twig');
-        }
-
-        $tokenEntity = new Entity\Tokens($token[0]);
-
-        // Get and empty user
-        $userEntity = new Entity\Users();
-
-        // Set the roles of the invitation code
-        $userEntity->setRoles($tokenEntity->getRoles());
-
-        // Get the form
-        $form = $this->getUserForm($userEntity, true);
-
-        // Set the validation
-        $form = $this->setUserFormValidation($form, true);
-
-        /** @var \Symfony\Component\Form\Form */
-        $form = $form->getForm();
-
-        // Check if the form was POST-ed, and valid. If so, store the user.
-        if ($request->isMethod('POST') && $response = $this->firstPost($request, $form)) {
-            return $response;
-        }
-
-        $context = [
-            'kind'        => 'create',
-            'form'        => $form->createView(),
-            'sitename'    => $this->getOption('general/sitename'),
-        ];
-
-        return $this->render('@bolt/invitation/invitation.twig', $context);
     }
 
     /**
@@ -601,7 +337,7 @@ class Users extends BackendBase
         foreach ($contenttypes as $contenttype) {
             foreach (array_keys($permissions) as $permission) {
                 $effectivePermissions[$contenttype['slug']][$permission] =
-                $this->app['permissions']->getRolesByContentTypePermission($permission, $contenttype['slug']);
+                    $this->app['permissions']->getRolesByContentTypePermission($permission, $contenttype['slug']);
             }
         }
         $globalPermissions = $this->app['permissions']->getGlobalRoles();
@@ -723,94 +459,6 @@ class Users extends BackendBase
                     'attr'        => ['placeholder' => Trans::__('page.edit-users.placeholder.displayname')],
                 ]
             )
-        ;
-
-        return $form;
-    }
-
-    /**
-     * Create a invitation generation form with the form builder.
-     *
-     * @return \Symfony\Component\Form\FormBuilder
-     */
-    private function getGenerateInvitationForm()
-    {
-        // Start building the form
-        $form = $this->createFormBuilder(FormType::class);
-
-        $roles = array_map(
-            function ($role) {
-                return $role['label'];
-            },
-            $this->app['permissions']->getDefinedRoles()
-        );
-
-        $form
-            ->add(
-                'roles',
-                ChoiceType::class,
-                [
-                    'choices'  => $roles,
-                    'expanded' => true,
-                    'multiple' => true,
-                    'label'    => Trans::__('page.invitation.label.assigned-roles'),
-                ]
-            );
-
-
-
-        return $form;
-    }
-
-    /**
-     * Create a email invitation form with the form builder.
-     *
-     *
-     * @return \Symfony\Component\Form\FormBuilder
-     */
-    private function getInvitationEmailForm()
-    {
-        // Start building the form
-        $form = $this->createFormBuilder(FormType::class);
-
-        $form
-            ->add(
-                'to',
-                TextType::class,
-                [
-                    'constraints' => new Assert\Email(),
-                    'label'       => Trans::__('page.invitation.share-options.to-email'),
-                    'attr'        => [
-                        'placeholder' => Trans::__('page.invitation.share-options.to-placeholder'),
-                        'class' => 'to',
-                    ],
-                ]
-            )
-            ->add(
-                'subject',
-                TextType::class,
-                [
-                    'constraints' => [new Assert\NotBlank(), new Assert\Length(['min' => 2, 'max' => 32])],
-                    'label'       => Trans::__('page.invitation.share-options.subject-email'),
-                    'attr'        => [
-                        'placeholder' => Trans::__('page.invitation.share-options.subject-placeholder'),
-                        'class' => 'subject',
-                    ],
-                ]
-            )
-            ->add(
-                'message',
-                TextareaType::class,
-                [
-                    'constraints' => [new Assert\NotBlank()],
-                    'label'       => Trans::__('page.invitation.share-options.message-email'),
-                    'attr'        => [
-                        'placeholder' => Trans::__('page.invitation.share-options.message-placeholder'),
-                        'class' => 'message',
-                    ],
-                ]
-            )
-
         ;
 
         return $form;
@@ -1085,4 +733,52 @@ class Users extends BackendBase
             $this->app['logger.system']->error("The 'mailoptions' need to be set in app/config/config.yml", ['event' => 'config']);
         }
     }
+
+    /**
+     * Create user using invitation code.
+     *
+     * @param String $code The invitation code provided
+     *
+     * @return \Bolt\Response\BoltResponse|\Symfony\Component\HttpFoundation\RedirectResponse
+     */
+    public function invitation(Request $request, $code)
+    {
+        // Check if the invitation code exists and it is still available
+        $token = $this->getRepository('Bolt\Storage\Entity\Tokens')->findBy(array('token' => $code));
+
+        if ((!$token) || (new \DateTime() > new \DateTime($token[0]->expiration))){
+        return $this->render('@bolt/invitation/invitationfail.twig');
+    }
+
+        $tokenEntity = new Entity\Tokens($token[0]);
+
+        // Get an empty user
+        $userEntity = new Entity\Users();
+
+        // Set the roles of the invitation code
+        $userEntity->setRoles($tokenEntity->getRoles());
+
+        // Get the form
+        $form = Users::getUserForm($userEntity, true);
+
+        // Set the validation
+        $form = $this->setUserFormValidation($form, true);
+
+        /** @var \Symfony\Component\Form\Form */
+        $form = $form->getForm();
+
+        // Check if the form was POST-ed, and valid. If so, store the user.
+        if ($request->isMethod('POST') && $response = $this->firstPost($request, $form)) {
+            return $response;
+        }
+
+        $context = [
+            'kind' => 'create',
+            'form' => $form->createView(),
+            'sitename' => $this->getOption('general/sitename'),
+        ];
+
+        return $this->render('@bolt/invitation/invitation.twig', $context);
+    }
+
 }
