+++ src/Provider/ControllerServiceProvider.php
@@ -77,6 +77,11 @@ class ControllerServiceProvider implements ServiceProviderInterface, EventSubscr
                 return new Controller\Backend\Users();
             }
         );
+        $app['controller.backend.invitation'] = $app->share(
+            function () {
+                return new Controller\Backend\Invitation();
+            }
+        );
 
         $app['controller.async.general'] = $app->share(
             function () {
@@ -161,6 +166,7 @@ class ControllerServiceProvider implements ServiceProviderInterface, EventSubscr
             'log',
             'records',
             'users',
+            'invitation',
         ];
         foreach ($backendKeys as $controller) {
             $event->mount($prefix, $app['controller.backend.' . $controller]);
